Move task to create spotify playlist to moodytunes app
