+++ apps/moodytunes/tasks.py
@@ -45,3 +45,36 @@ def fetch_song_from_spotify(self, spotify_code, username='anonymous'):
             logger.info('{} - Created song {} in database'.format(signature, spotify_code))
         else:
             logger.info('{} - Did not create song {} in database, song already exists'.format(signature, spotify_code))
+
+
+@task(bind=True, max_retries=3, default_retry_delay=60*15)
+def create_spotify_playlist_from_songs(self, auth_code, spotify_user_id, playlist_name, songs):
+    """
+    Create a playlist on a user's Spotify account from a list of tracks in our system.
+
+    :param auth_code: (str) SpotifyUserAuth access_token for the given user
+    :param spotify_user_id: (str) Spotify username for the given user
+    :param playlist_name: (str) Name of the playlist to be created
+    :param songs: (list) Collection of Spotify track URIs to add to playlist
+
+    """
+    spotify = SpotifyClient()
+
+    try:
+        logger.info('Creating playlist for user {}'.format(spotify_user_id))
+        playlist_id = spotify.create_playlist(auth_code, spotify_user_id, playlist_name)
+        logger.info('Created playlist for user {} successfully'.format(spotify_user_id))
+    except SpotifyException:
+        logger.warning('Error creating playlist for user {}, retrying...'.format(spotify_user_id))
+        self.retry()
+
+    try:
+        logger.info('Adding songs to playlist {}'.format(playlist_id))
+        spotify.add_songs_to_playlist(auth_code, playlist_id, songs)
+        logger.info('Added songs to playlist {} successfully'.format(playlist_id))
+    except SpotifyException:
+        logger.warning('Error adding songs to playlist {} for user {}, retrying...'.format(
+            playlist_id,
+            spotify_user_id
+        ))
+        self.retry()
